requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
faiss-cpu>=1.7.4
ftfy>=6.1.3
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
requests>=2.31.0
huggingface_hub>=0.20.0
datasets>=2.18.0
open_clip_torch>=2.24.0
sentence-transformers>=2.7.0
einops>=0.7.0
